Decompile func0f11416c

+++ src/include/game/game_113220.h
@@ -12,6 +12,6 @@ u32 func0f1138a4(void);
 u32 func0f113f10(void);
 u32 func0f1140bc(void);
 void func0f11412c(f32 speed);
-u32 func0f11416c(void);
+f32 func0f11416c(void);
 
 #endif
